+++ moodterm/moodtermApp.swift
@@ -1,0 +1,17 @@
+//
+//  moodtermApp.swift
+//  moodterm
+//
+//  Created by Emma Puls on 27/10/2024.
+//
+
+import SwiftUI
+
+@main
+struct moodtermApp: App {
+    var body: some Scene {
+        WindowGroup {
+            ContentView()
+        }
+    }
+}