+++ CHANGELOG.md
@@ -1,15 +1,26 @@
 # moodterm
 
+## 0.3.0
+
+### Minor Changes
+
+#### Switched to SwiftUI & Swift
+And I have a terminal GUI! It works! 
+
+##### Issues
+- Text won't scroll
+- Not handling ANSI colours
+
 ## 0.2.0
 
 ### Minor Changes
 
-- ### Context
+- #### Context
   - I've completed the ownership and structs sections of [The Rust Programming Language](https://doc.rust-lang.org/book/title-page.html)
   - I've been reviewing the [cacao examples](https://github.com/ryanmcgrath/cacao) to understand how it works better
     - I've spent a bit of time understanding why NSTextView is not yet supported
     - I've decided to continue using cacao for now for the UI
   - I've been reading about concurrency
   
-  ### What is in this change
+  #### What is in this change
   - I've refactored the UI a bit to seperate the concerns and done a bit of a visual refresh